+++ PointOfSale/Models/VMGastos.cs
@@ -19,7 +19,7 @@ namespace PointOfSale.Models
         public DateTime? ModificationDate { get; set; }
         public string? ModificationUser { get; set; }
         public decimal? ImporteSinIva { get; set; }
-        public int? Iva { get; set; }
+        public decimal? Iva { get; set; }
         public decimal? IvaImporte { get; set; }
         public string? NroFactura { get; set; }
         public string? TipoFactura { get; set; }
