+++ PointOfSale/Models/VMProveedorMovimiento.cs
@@ -7,7 +7,7 @@ namespace PointOfSale.Models
         public int IdProveedorMovimiento { get; set; }
         public decimal Importe { get; set; }
         public decimal? ImporteSinIva { get; set; }
-        public int? Iva { get; set; }
+        public decimal? Iva { get; set; }
         public decimal? IvaImporte { get; set; }
         public string? NroFactura { get; set; }
         public string? TipoFactura { get; set; }
